Fixed UNDO log truncation.
So far the truncation did not really adjust the log size, so another session
could write records into a log which was not the last in the sequence.

This commit removes the recent_end field from the UndoRecordSet because we'd
also need to cache the log size and that would make the logic non-trivial.