Handle failed undo of aborted subtransaction.
If the top-level transaction is being aborted, the undo request (created also
in the case of foreground undo) gets prepared. If the undo then fails, the
backend tries to reschedule it for background undo before it
terminates. However the request is not prepared if only a subtransaction is
aborted. This lead to an assertion failure in RescheduleUndoRequest().

Since the problem occurred in proc_exit() callback, we probably should not try
to reschedule the request in such a case at all.