Fixed off-by one issue when checking the "discard" pointer.
So far UndoLogRecPtrIsDiscardedSlowPath() wasn't consistent with the
definition that "discard" is the "oldest data needed" - instead, it behaved as
if even (pointer - 1) was needed. To compensate for this inaccuracy,
undofile_read() had to pass to UndoRecPtrIsDiscarded() the 0-th byte of the
next block rather than the last byte of the current block. As a consequence,
UndoRecPtrIsDiscarded() used the wrong cache entry if the pointer to be
checked was the last byte of the current log.