Avoid executing undo actions multiple times.
This is an attempt to move the checks whether particular undo action was
already executed from the AM (zheap) to the undo layer. That might simplify
the AM a bit, and the it seems more appropriate for the undo subsystem to take
care.

The WAL records generated during undo execution carry the undo record pointer,
so if the less than the whole URS is executed due to server crash, we can
recognize during recovery which undo records should not be executed again.