Recover undo requests from the undo files instead from checkpoint.
Until this commit, checkpoint included undo requests, so the requests that
could not be processed due to server crash can be recovered during crash
recovery. This seems to make some recovery scenarios tricky. One particular
example could be seen in CloseDanglingUndoRecordSets(). If the undo record set
is closed, the corresponding WAL is written, checkpoint is performed and the
server crashes before the undo request could be created for the now-closed
set, the undo request will not exist after crash recovery.

I'm not sure if the checkpoint can really take place at the stage described
here, but the potential problems that the checkpoint timing might cause at
least make thinking about recovery harder. In general, it seems to be easier
to maintain consistency if we store status information at fewer places.

With this commit in place, we scan the undo logs at the end of recovery in
order to retrieve the information needed to register the undo requests, and
register those requests before the undo worker is allowed to start.

The commit also adds last_rec_applied field to UndoRecordSetChunkHeader, which
should help us to avoid repeated execution of particular undo record - it
seems more suitable for the undo infrastructure to take this responsibility so
that the zheap access method does not have to care. This tracking feature was
actually the reason for the removal of undo request metadata from the
checkpoint (the previous idea was to store the last_rec_applied pointer in the
undo request), but could have been done in a separate commit.