Cleaning up orphaned files using undo logs
This is based on the patch last version of which was published here:

https://www.postgresql.org/message-id/87363.1611941415%40antos

In addition, views pg_stat_undo_chunks and pg_stat_undo_records have been
added to make the pg_undo_dump-like functionality available in backend. These
views (and the underlying functions) should make the future development easier
because they read the information directly from the shared memory. In
contrast, the pg_undo_dump utility reads the data from disk so when we need to
see the most recent changes, a checkpoint has to be enforced in most cases.