Introduce undo worker to perform cleanup after crash recovery.
Altough we currently do not support background UNDO as such, a background
worker seems to be the easiest solution to apply the undo records of
transactions that could not complete due to server crash. The startup process
cannot perform this task because database connections and transactions are
needed to lock relations and to access the system catalogs.

As soon as crash recovery is done, one undo worker is launched for each
database. Once the pending undo is applied, those workers stop and do not
start again.