Adjusted logical decoding so it works with the zheap AM.
This is the first step to make the logical replication work with zheap. The
next step will be to implement the scan using the "historic" MVCC snapshot so
that the initial load of tables can be performed.

Quite a few bugs fixed in order to make the regression tests
(contrib/test_decoding) work.