Set the snapshot so that UPDATE and DELETE changes can be applied.
NULL snapshot worked for the heap AM because neither heapam_tuple_update()
nore heapam_tuple_delete() use the snapshot.

Although the 84b137bd commit message says that we still need to implement scan
using "historic MVCC" snapshot, this is actually not needed because the
logical replication worker converts the historic snapshot to the regular MVCC
snapshot before it uses it for the initial load, see
SnapBuildInitialSnapshot(). So the logical replication for zheap should work
now.